move delta and truth table to separate mod
[ci skip]
